+++ qtodotxt/ui/views/tasks_search_view.py
